+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- transfer.scheduled parameter to allow Transfer scheduling
+- StarkBank.Transfer.delete to cancel scheduled Transfers
 
 ## [0.5.1] - 2020-06-09
 ### Fixed